requirements.txt
Django==4.0.2
mysqlclient==2.1.0
djangorestframework==3.13.1
black==22.1.0